+++ tomcat/src/main/java/org/apache/coyote/http11/request/Path.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.Map;
