+++ tomcat/src/test/java/org/apache/coyote/http11/ExtensionTest.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
+import org.apache.coyote.http11.request.Extension;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
