+++ tomcat/src/test/java/org/apache/coyote/http11/PathTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.coyote.http11.request.Path;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +15,7 @@ class PathTest {
     @Test
     void from() {
         final String uri = "/index.html";
-        final String expectedResource = "/index.html";
+        final String expectedResource = "/index";
         final String expectedContentType = "text/html";
 
         Path path = Path.from(uri);
@@ -30,7 +31,7 @@ class PathTest {
     @Test
     void from_With_QueryParam() {
         final String uri = "/index.html?username=east&password=password";
-        final String expectedResource = "/index.html";
+        final String expectedResource = "/index";
         final String expectedContentType = "text/html";
         final Map<String, String> expectedParams = new HashMap<>();
         expectedParams.put("username", "east");
